+++ lib/src/custom_layout.dart
@@ -198,19 +198,21 @@ abstract class _CustomLayoutStateBase<T extends _SubSwiper> extends State<T>
     if (_lockScroll) return;
 
     final velocity = widget.scrollDirection == Axis.horizontal
-        ? details.velocity.pixelsPerSecond.dx
+        ? widget.axisDirection == AxisDirection.left
+            ? details.velocity.pixelsPerSecond.dx
+            : -details.velocity.pixelsPerSecond.dx
         : details.velocity.pixelsPerSecond.dy;
 
     if (_animationController.value >= 0.75 || velocity > 500.0) {
       if (_currentIndex <= 0 && !widget.loop) {
         return;
       }
-      _move(1.0, nextIndex: _currentIndex - 1);
+      _move(1.0, nextIndex: _currentIndex + 1);
     } else if (_animationController.value < 0.25 || velocity < -500.0) {
       if (_currentIndex >= widget.itemCount - 1 && !widget.loop) {
         return;
       }
-      _move(0.0, nextIndex: _currentIndex + 1);
+      _move(0.0, nextIndex: _currentIndex - 1);
     } else {
       _move(0.5);
     }
@@ -220,7 +222,9 @@ abstract class _CustomLayoutStateBase<T extends _SubSwiper> extends State<T>
     if (_lockScroll) return;
     _currentValue = _animationController.value;
     _currentPos = widget.scrollDirection == Axis.horizontal
-        ? details.globalPosition.dx
+        ? widget.axisDirection == AxisDirection.left
+            ? details.globalPosition.dx
+            : -details.globalPosition.dx
         : details.globalPosition.dy;
   }
 
@@ -228,7 +232,9 @@ abstract class _CustomLayoutStateBase<T extends _SubSwiper> extends State<T>
     if (_lockScroll) return;
     var value = _currentValue +
         ((widget.scrollDirection == Axis.horizontal
-                    ? details.globalPosition.dx
+                    ? widget.axisDirection == AxisDirection.left
+                        ? details.globalPosition.dx
+                        : -details.globalPosition.dx
                     : details.globalPosition.dy) -
                 _currentPos) /
             _swiperWidth /