+++ CHANGELOG.md
@@ -1,37 +1,68 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
+
 ### Changed
+
 ### Deprecated
+
 ### Removed
+
 ### Fixed
+
 ### Security
 
+## [2.0.1] - 2022-01-11
+
+### Added
+
+- [Issue#3](https://github.com/TheAnkurPanchani/card_swiper/issues/3): Support `AxisDirection` left and right for Stack layout
+
+### Fixed
+
+- Fix pan event index update for custom layout
+
+## [2.0.0] - 2022-01-09
+
+### Changed
+
+- [PR#17](https://github.com/TheAnkurPanchani/card_swiper/pull/17): Migrate from `pedantic` to `flutter_lints_plus` and fix problems related to lint
 
 ## [1.0.4] - 2021-08-06
+
 ### Fixed
+
 - Initial index should set from controller is provided in parameter
 - Update readme
 
 ## [1.0.3] - 2021-08-06
+
 ### Fixed
+
 - Initial index should set from controller is provided in parameter
 
 ## [1.0.2] - 2017-06-20
+
 ### Fixed
+
 - README text and references
 - Library name
 - Example reference
 - Format according to dartfmt
 
 ## [1.0.1] - 2017-06-20
+
 ### Fixed
+
 - Format according to dartfmt
 
 ## [1.0.0] - 2017-06-20
+
 - First release after fork